OpenFoodFact-G2-main/back/backoffice/user/serializers.py
from rest_framework import serializers
from .models import Product

class ProductSerializer(serializers.ModelSerializer):
    user = serializers.PrimaryKeyRelatedField(read_only=True, default=serializers.CurrentUserDefault())
    class Meta:
        model = Product
        fields = ['id', 'user', 'data']  # Incluez le champ 'data' dans les champs du serializer
